☘️ refactor [#11.6.1]: 1차 개선 - Admin 라우트 Route Group 리팩토링 및 다국어 렌더링 병합
- [Architecture] 기존 `app/[locale]` 하위의 유저 페이지들을 `(main)` 라우트 그룹으로 캡슐화하여 의존성 분리 완료
- [Bug Fix] `admin/layout.tsx`에서 발생할 수 있는 `Nested HTML/Body` 태그 충돌 버그 해결 (최상위 Root Layout으로 통합)
- [i18n] 미들웨어 정규식을 롤백하여 Admin 영역도 동일한 다국어 파이프라인을 공유하도록 개선 (하드코딩 제거)
- [Clean Code] 폰트 및 NextIntl 글로벌 Provider 선언을 최상단 루트 하나로 통합하여 코드 중복 제거

🔗 Related:
- Issue [#866]
- ai-bot-review (https://github.com/jjaayy2222/flownote-mvp/pull/869#pullrequestreview-4026951114)
