+++ web_ui/src/app/[locale]/(admin)/admin/analytics/page.tsx
