+++ web_ui/src/app/[locale]/(main)/chat/page.tsx
