+++ web_ui/src/app/[locale]/(main)/graph/page.tsx
