+++ web_ui/src/app/[locale]/(main)/page.tsx
