+++ web_ui/src/app/[locale]/(main)/search/page.tsx
