+++ web_ui/src/app/[locale]/(main)/stats/page.tsx
